ensure we don't use bootconsoles after init has been released
This is a followup to the cleanups for earlyprintk patch from Gerd Hoffmann
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commit;h=69331af79cf29e26d1231152a172a1a10c2df511

and has been committed to the -mm tree. Its filename is
ensure-we-dont-use-bootconsoles-after-init-has-been-released.patch

This ensures that a bootconsole is unregistered if it is not replaced.
The current implementation spews garbage out the bootconsole in this
case, since the bootconsole structure is normally in the init section, and is
freed, but still used.

Acked-by: Gerd Hoffmann <kraxel@redhat.com>
Acked-by: Paul Mundt <lethal@linux-sh.org>



git-svn-id: svn://localhost/svn/linux-kernel/trunk@3589 526b6c2d-f592-4532-a319-5dd88ccb003d