[!no_src_qa!] Task[#2424] Kernel updating to 2.6.22.4
We (the -stable team) are announcing the release of the 2.6.22.4 kernel.
It contains one security fix and all users of the 2.6.22 series are
encouraged to upgrade to it.

I'll also be replying to this message with a copy of the patch between
2.6.22.3 and 2.6.22.4

The updated 2.6.22.y git tree can be found at:
       git://git.kernel.org/pub/scm/linux/kernel/git/stable/linux-2.6.22.y.git
and can be browsed at the normal kernel.org git web browser:
       http://git.kernel.org/?p=linux/kernel/git/stable/linux-2.6.22.y.git;a=summary

thanks,

greg k-h

--------

 Makefile  |    2 +-
 fs/exec.c |   13 +++++++++----
 2 files changed, 10 insertions(+), 5 deletions(-)

Summary of changes from v2.6.22.3 to v2.6.22.4
==============================================

Greg Kroah-Hartman (1):
     Linux 2.6.22.4

Marcel Holtmann (1):
     Reset current->pdeath_signal on SUID binary execution (CVE-2007-3848)



git-svn-id: svn://localhost/svn/linux-kernel/trunk@3592 526b6c2d-f592-4532-a319-5dd88ccb003d