[#3594] [ltp] fix bug in test case readv02 failed on all boards
In Bernds' r3704 commit:
Fix a problem reported in the forums - libstdc++ can call writev with an
iovec containing { NULL, 0 }, which works fine on i686-linux, but fails on
Blackfin.  Fixed by allowing size 0 transfers to/from userspace regardless
of the address.

The writev bug was fixed, but introduced the readv bug here.

Fix it by testing readv or writev in access_ok



git-svn-id: svn://localhost/svn/linux-kernel/trunk@3804 526b6c2d-f592-4532-a319-5dd88ccb003d