[#2424] Upgrade kernel from 2.6.22.12 to 2.6.22.13
We (the -stable team) are announcing the release of the 2.6.22.13 kernel.
If fixes some security issues and any user of the 2.6.22 series is
encouraged to upgrade.

I'll also be replying to this message with a copy of the patch between
2.6.22.12 and 2.6.22.13

The updated 2.6.22.y git tree can be found at:
       git://git.kernel.org/pub/scm/linux/kernel/git/stable/linux-2.6.22.y.git
and can be browsed at the normal kernel.org git web browser:
       http://git.kernel.org/?p=linux/kernel/git/stable/linux-2.6.22.y.git;a=summary

thanks,

greg k-h

--------

 Makefile             |    2 +-
 kernel/exit.c        |    3 +--
 net/ipv4/tcp_input.c |    5 +++++
 3 files changed, 7 insertions(+), 3 deletions(-)

Summary of changes from v2.6.22.12 to v2.6.22.13
================================================

Greg Kroah-Hartman (1):
     Linux 2.6.22.13

Ilpo Järvinen (1):
     TCP: Make sure write_queue_from does not begin with NULL ptr (CVE-2007-5501)

Roland McGrath (1):
     wait_task_stopped: Check p->exit_state instead of TASK_TRACED (CVE-2007-5500)



git-svn-id: svn://localhost/svn/linux-kernel/trunk@3905 526b6c2d-f592-4532-a319-5dd88ccb003d