[#3719] When the CPLBs get a miss, we do:   - find a victim in the HW table   - remove the victim   - find the replacement in the software table   - put it into the HW table.
If we can't find a replacement in the software table, we accidently 
leave a duplicate in the HW table. This patch ensures that duplicate
is marked as not valid.

What we should do is find the replacement in the software table, before
we find a victim in the HW table - but its too late in the release cycle
to do that much restructuring of this code.



git-svn-id: svn://localhost/svn/linux-kernel/trunk@3926 526b6c2d-f592-4532-a319-5dd88ccb003d