Fix BUG [#3627] bf548 SMSC911x/921x driver fails to call netpoll callback in kgdb at the beginning of kernel boot up.
* Sync local driver with latest version of the smsc911x driver posted to
* the netdev list

Bahadir has recently been modifying this driver for the forthcoming NAPI
changes in 2.6.24,
but I don't think this has been tested yet.

-Michael 

http://marc.info/?l=linux-netdev&m=118461459010118&w=4

Attached is a driver for SMSC's LAN911x and LAN921x families of embedded
ethernet controllers.

There is an existing smc911x driver in the tree; this is intended to
replace it.  Dustin McIntire (the author of the smc911x driver) has
expressed his support for switching to this driver.

This driver contains workarounds for all known hardware issues, and has
been tested on all flavours of the chip on multiple architectures.

Signed-off-by: Steve Glendinning <steve.glendinning@smsc.com>
Signed-off-by: Bahadir Balban <Bahadir.Balban@arm.com>
Signed-off-by: Dustin Mcintire <dustin@sensoria.com>
Signed-off-by: Bill Gatliff <bgat@billgatliff.com>



git-svn-id: svn://localhost/svn/linux-kernel/trunk@4047 526b6c2d-f592-4532-a319-5dd88ccb003d