Task[#2326] This fixes a crash on bootup with CONFIG_MPU on the BF548.
The function flush_switched_dcplbs was clearing the CPLB entries covering
the process permission bitmasks.  This means that the sequence

	flush_switched_dcplbs ();
        set_mask_dcplbs(mm->context.page_rwx_mask);

has a problem: if kernel code (such as an interrupt) causes a CPLB miss before
set_mask_dcplbs completes, the CPLB handler function causes a double fault,
with an instantaneous reboot.


This bug fix is dedicated to Michael Hennerich, the only person in the world
capable of providing working JTAG hardware.



git-svn-id: svn://localhost/svn/linux-kernel/trunk@4539 526b6c2d-f592-4532-a319-5dd88ccb003d