[#2424] Prevent potential Core Hang situation: Add inlines for PLL_CTL
If the new value written to the PLL_CTL or VR_CTL register is the
same as the previous value, the PLL wake-up will occur immediately
(PLL is already locked), but the core and system clock will be
bypassed for the PLL_LOCKCNT duration. For this interval, code will
execute at the CLKIN rate instead of at the expected CCLK rate.
Software should guard against this condition by comparing the
current value to the new value before writing the new value.




git-svn-id: svn://localhost/svn/linux-kernel/trunk@4598 526b6c2d-f592-4532-a319-5dd88ccb003d