Task: [#3862] Functional power management support
Blacklist PM_SUSPEND_MEM for BF533/2/1:
On BF533/2/1:
If we enter Hibernate the SCKE Pin is driven Low,
so that the SDRAM enters Self Refresh Mode.
However when the reset sequence that follows hibernate
state is executed, SCKE is driven High, taking the
SDRAM out of Self Refresh.

If you reconfigure and access the SDRAM "very quickly",
you are likely to avoid errors, otherwise the SDRAM
start losing its contents.
An external HW workaround is possible using logic gates.




git-svn-id: svn://localhost/svn/linux-kernel/trunk@5007 526b6c2d-f592-4532-a319-5dd88ccb003d