Fix a few problems I discovered when building a kernel with upstream CVS binutils.
We have to add the NOTES macro to our linker script, since a kernel 
built with --build-id is otherwise unable to boot.  Last time NOTES was
added, it broke things, but the definition of the macro has changed not
to rely on parts of the linker script that aren't present on Blackfin.

I also noticed that _l2_lma_start does not point into the kernel image, but
rather somewhere in L1/L2 space, which seems unintended.  Also, when the L2
section was added to the linker script, the part following it which computes
then length of the init section was not updated.



git-svn-id: svn://localhost/svn/linux-kernel/trunk@5758 526b6c2d-f592-4532-a319-5dd88ccb003d