This is the result that came out of patch [#4677].  It is a mixture of Michael McTernan's patch and the existing cplb-mpu code.
We ditch the old cplb-nompu implementation, which is a good example of why
a good algorithm in a HLL is preferrable to a bad algorithm written in
assembly.  Rather than try to construct a table of all posible CPLBs and
search it, we just create a (smaller) table of memory regions and
their attributes.  Some of the data structures are now unified for both the
mpu and nompu cases.  A lot of needless complexity in cplbinit.c is removed.

Further optimizations:
  * compile cplbmgr.c with a lot of -ffixed-reg options, and omit saving
    these registers on the stack when entering a CPLB exception.
  * lose cli/nop/nop/sti sequences for some workarounds - these don't make
    sense in an exception context

Additional code unification should be possible after this.



git-svn-id: svn://localhost/svn/linux-kernel/trunk@5856 526b6c2d-f592-4532-a319-5dd88ccb003d