[!no_src_qa!]Fix bug[#4914] Enable hardware based CTS and RTS for bf548.
Both software emulated and hardware based CTS and RTS are enabled in
serial driver.

The CTS RTS PIN connection on BF548 UART port is defined as a modem
device not as a host device. In order to test it under Linux, please
nake a cross UART cable to exchange CTS and RTS signal.



git-svn-id: svn://localhost/svn/linux-kernel/trunk@6140 526b6c2d-f592-4532-a319-5dd88ccb003d