bug [#5647] - check in patch by Eduardo Tagle <ejtagle@tutopia.com>
When the bus owner unlocks the spi bus, there may be pending messages on the queue from other spi devices. The messages will stay on the queue until a spi device requests another transfer and starts pumping the message queue. This patch schedules messages transfer when spi bus is released.


git-svn-id: svn://localhost/svn/linux-kernel/branches/2009R1@7750 526b6c2d-f592-4532-a319-5dd88ccb003d